+++ packages/iti-react/src/components/dialog/EasyFormDialog.tsx
@@ -91,9 +91,8 @@ export interface EasyFormDialogProps {
 
     /**
      * A callback that fires when the dialog has completely closed. Your
-     * `onClose` callback should call, for example,
-     * `setDialogVisible(false)` so that the `EasyFormDialog` is no longer
-     * rendered.
+     * `onClose` callback should call, for example, `setDialogVisible(false)` so
+     * that the `EasyFormDialog` is no longer rendered.
      */
     onClose(): void